step 7 리팩터링 : interface 상속받도록 수정
