+++ src/main/java/roomescape/SpringConfig.java
@@ -1,9 +1,8 @@
 package roomescape;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.controller.ReservationController;
-import roomescape.repository.ReservationRepository;
+import roomescape.repository.JdbcReservationRepository;
 
 import javax.sql.DataSource;
 
@@ -21,7 +20,7 @@ public class SpringConfig {
     }
 
     @Bean
-    public ReservationRepository reservationRepository() {
-        return new ReservationRepository(dataSource);
+    public JdbcReservationRepository reservationRepository() {
+        return new JdbcReservationRepository(dataSource);
     }
 }
