+++ src/main/java/roomescape/domain/dao/ReservationDAO.java
@@ -1,0 +1,4 @@
+package roomescape.domain.dao;
+
+public class ReservationDAO {
+}
